# beatstore/src/app/generated/model/beat.ts
/**
 * Beat store API
 * Beat store RS API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Beat { 
    guid?: string;
    name?: string;
    beatimg?: Blob;
    price?: number;
    BPM?: number;
    likeNumber?: number;
    genre?: string;
    mp3ID?: string;
    producedby?: string;
    howmanysold?: number;
    bPM?: number;
}