# beatstore/src/app/generated/model/beatMP3FullDTO.ts
/**
 * Beat store API
 * Beat store RS API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BeatMP3FullDTO { 
    guid?: string;
    beatid?: string;
    fullbeatmp3?: Blob;
}