# beatstore/src/app/generated/model/userDTO.ts
/**
 * Beat store API
 * Beat store RS API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Beat } from './beat';


export interface UserDTO { 
    guid?: string;
    name?: string;
    bio?: string;
    email?: string;
    firebase_id?: string;
    profilepicture?: Blob;
    likedbeats?: Array<Beat>;
    beatsincart?: Array<Beat>;
    boughtbeats?: Array<Beat>;
    soldbeats?: Array<Beat>;
    bankaccount?: number;
    paypalmail?: string;
}